Fixes for AMDGPU calling convention (#524)
AMDGPU backend requires the use of `C_AGGREGATE_REG` (i.e., exploding
a struct into its component fields) calling convention rather than
`C_AGGREGATE_MEM` (i.e., using `byval` and `sret` to pass the struct
on the stack). This change checks the target machine and generates the
correct code in the AMDGPU case.

Arguably, this was intended to be done within `CCallingConv` itself,
but that would be a larger change and would also involve potentially a
lot of code duplication. I do not attempt to do that in this change.

I also modified `C_AGGREGATE_REG` to correctly handle nested structs,
which otherwise are not properly exploded. (The previous code
unwrapped only one level of the struct.)

Contains one other typo fix for aggregate stores to make sure we use
the right address space.